Selenium_Ecommerce/requirements.txt
selenium==4.36.0
pytest==8.4.2
pytest-xdist==3.8.0
allure-pytest==2.15.0 # For Excel
pytest-html==4.1.1
webdriver-manager==4.0.2
openpyxl==3.1.5
pytest-rerunfailures==16.1